requirements.txt
annotated-types==0.7.0
coverage==7.5.1
gitdb==4.0.11
GitPython==3.1.43
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
pluggy==1.5.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pytest==8.2.1
pytest-cov==5.0.0
python-dotenv==1.0.1
rich==13.7.1
smmap==5.0.1
typing_extensions==4.11.0
